// frontend/src/app/services/api/models/heating-schedule.ts
export interface HeatingSchedule {
  day: WeekDay;
  schedule: {
'00:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'00:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'01:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'01:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'02:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'02:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'03:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'03:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'04:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'04:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'05:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'05:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'06:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'06:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'07:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'07:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'08:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'08:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'09:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'09:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'10:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'10:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'11:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'11:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'12:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'12:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'13:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'13:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'14:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'14:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'15:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'15:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'16:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'16:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'17:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'17:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'18:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'18:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'19:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'19:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'20:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'20:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'21:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'21:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'22:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'22:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'23:00'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
'23:30'?: 'COMFORT' | 'COMFORT_1' | 'COMFORT_2' | 'ECO' | 'FROST_PROTECTION' | 'OFF' | 'UNKNOWN';
};
}